fix output

+++ source/Lib/vvenc/vvencCfg.cpp
@@ -3067,7 +3067,7 @@ VVENC_DECL const char* vvenc_get_config_as_string( vvenc_config *c, vvencMsgLeve
       css << "QuantThr: " << (c->m_quantThresholdVal >> 1) << ".5 ";
     else
       css << "QuantThr: " << (c->m_quantThresholdVal >> 1) << " ";
-    css << "SelectiveRDQO: " << c->m_useSelectiveRDOQ << " ";
+    css << "SelectiveRDQO: " << ( int ) c->m_useSelectiveRDOQ << " ";
 
     css << "\nRATE CONTROL CFG: ";
     css << "RateControl:" << ( c->m_RCTargetBitrate > 0 ) << " ";
